Refactor httpClientRequests to add responseTimeout #deploy-test-dolly-backend #deploy-dolly-backend
Implemented a standard timeout defined as REQUEST_DURATION constant used across multiple classes' httpClientRequests. This refactoring ensures the application aborts requests to prevent hang-ups due to slow or unresponsive servers.
