+++ apps/dolly-backend/src/main/java/no/nav/dolly/bestilling/pensjonforvalter/PensjonforvalterConsumer.java
@@ -41,7 +41,6 @@ import static no.nav.dolly.util.JacksonExchangeStrategyUtil.getJacksonStrategy;
 @Service
 public class PensjonforvalterConsumer implements ConsumerStatus {
 
-    public static final int REQUEST_DURATION = 30000;
     private final TokenExchange tokenService;
     private final WebClient webClient;
     private final ServerProperties serverProperties;
