+++ apps/dolly-backend/src/main/java/no/nav/dolly/util/RequestTimeout.java
@@ -1,0 +1,9 @@
+package no.nav.dolly.util;
+
+import lombok.experimental.UtilityClass;
+
+@UtilityClass
+public class RequestTimeout {
+
+    public static final int REQUEST_DURATION = 30000;
+}
